requirements.txt
lxml
pyyaml
svgpathtools
watchdog
